+++ src/provider/caspian-store-provider.tsx
@@ -14,6 +14,7 @@ import { AuthProvider } from '../context/auth-context';
 import { CartProvider } from '../context/cart-context';
 import { ScriptSettingsProvider } from '../context/script-settings-context';
 import { ThemeInjector } from '../context/theme-context';
+import { FontLoader } from '../context/font-loader';
 import { ToastProvider } from '../ui/toast';
 import { LocaleProvider } from '../i18n/locale-context';
 import type { MessageDict } from '../i18n/messages';
@@ -86,6 +87,7 @@ export function CaspianStoreProvider({
             <CartProvider db={value.firebase.db}>
               <ScriptSettingsProvider collections={value.collections}>
                 <ThemeInjector />
+                <FontLoader />
                 {children}
               </ScriptSettingsProvider>
             </CartProvider>